rename: 폴더명 및 폴더구조 변경
